Move summary statistic tests outside
